pipelineIQ/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
motor>=3.5.0,<3.7.0
beanie>=1.26.0,<1.28.0
python-jose[cryptography]>=3.4.0
httpx>=0.28.0
pydantic-settings>=2.8.0
python-dotenv>=1.1.0
